--- src/utils/renderDayChart.ts
@@ -1,201 +1,0 @@
-import * as d3 from 'd3';
-import { toISOString } from '@/utils/dateUtils';
-import * as constants from '../constants/dayChartConstants';
-import { ExtendedCandle, RefinedCandle } from './refineCandle';
-
-// 날짜 데이터 파싱 함수
-const parseDataWithDates = (refinedData: RefinedCandle[]): ExtendedCandle[] => {
-  // 날짜 기준 오름 차순 처리
-  const sortedData = [...refinedData].sort((a, b) => {
-    const dateA = new Date(a.timestamp).getTime();
-    const dateB = new Date(b.timestamp).getTime();
-    return dateA - dateB;
-  });
-
-  // dates 형식 날짜 기준으로 파싱
-  let prevMonth = 0;
-  return sortedData.map((d) => {
-    const date = new Date(d.timestamp);
-    const dateStr = toISOString(d.timestamp).split('T')[0];
-    const [, month] = dateStr.split('-').map(Number);
-    const isXAxisMark = prevMonth !== month;
-    prevMonth = month;
-
-    return {
-      ...d,
-      date,
-      dateStr,
-      isXAxisMark,
-    };
-  });
-};
-
-// 차트 스케일 생성 함수
-const createChartScales = (
-  parsedData: ExtendedCandle[],
-  width: number,
-  height: number,
-  margin: { top: number; right: number; bottom: number; left: number },
-  volumeHeight: number,
-) => {
-  const x = d3
-    .scaleBand()
-    .domain(parsedData.map((d) => d.dateStr))
-    .range([margin.left, width - margin.right])
-    .padding(0.1);
-
-  const minValue = d3.min(parsedData, (d) => d.low) ?? 0;
-  const maxValue = d3.max(parsedData, (d) => d.high) ?? 1;
-
-  const y = d3
-    .scaleLinear()
-    .domain([minValue * 0.999, maxValue * 1.001])
-    .range([height - margin.bottom - volumeHeight, margin.top]);
-
-  // 거래량 y축
-  const volumeMax = d3.max(parsedData, (d) => d.volume) ?? 0;
-  const yVolume = d3
-    .scaleLinear()
-    .domain([0, volumeMax * 1.1])
-    .range([height - margin.bottom, height - margin.bottom - volumeHeight]);
-
-  return { x, y, yVolume };
-};
-
-// 축 생성 및 그리기 함수
-const drawAxes = (
-  parsedData: ExtendedCandle[],
-  x: d3.ScaleBand<string>,
-  yPrice: d3.ScaleLinear<number, number>,
-  yVolume: d3.ScaleLinear<number, number>,
-  gx: SVGGElement | null,
-  gy: SVGGElement | null,
-  width: number,
-  height: number,
-  margin: { top: number; right: number; bottom: number; left: number },
-) => {
-  if (!gx || !gy) return;
-
-  // x축
-  const xAxis = d3
-    .axisBottom(x)
-    .tickValues(parsedData.filter((d) => d.isXAxisMark).map((d) => d.dateStr))
-    .tickFormat((d) => {
-      const [, month] = d.split('-');
-      const monthIndex = parseInt(month) - 1;
-      return `${constants.MONTH_NAME_ENG[monthIndex]}`;
-    });
-
-  d3.select(gx)
-    .attr('transform', `translate(0, ${height - margin.bottom})`)
-    .call(xAxis);
-
-  // gy 내부 기존 축 그룹 초기화
-  d3.select(gy).selectAll('*').remove();
-
-  // 왼쪽: 거래량 y축
-  d3.select(gy)
-    .append('g')
-    .attr('class', 'y-axis-volume')
-    .attr('transform', `translate(${margin.left}, 0)`)
-    .call(d3.axisLeft(yVolume).ticks(3).tickFormat(d3.format('.2s')));
-
-  // 오른쪽: 시세 y축
-  d3.select(gy)
-    .append('g')
-    .attr('class', 'y-axis-price')
-    .attr('transform', `translate(${width - margin.right}, 0)`)
-    .call(d3.axisRight(yPrice).ticks(5).tickFormat(d3.format('~s')));
-};
-
-//거래량 봉 그래프 그리기 함수
-const drawVolumeBars = (
-  svg: d3.Selection<SVGSVGElement, unknown, null, undefined>,
-  parsedData: ExtendedCandle[],
-  x: d3.ScaleBand<string>,
-  yVolume: d3.ScaleLinear<number, number>,
-) => {
-  svg.selectAll('.volume-bar').remove();
-
-  svg
-    .selectAll('.volume-bar')
-    .data(parsedData)
-    .join('rect')
-    .attr('class', 'volume-bar')
-    .attr('x', (d) => x(d.dateStr) ?? 0)
-    .attr('y', (d) => yVolume(d.volume))
-    .attr('width', x.bandwidth())
-    .attr('height', (d) => yVolume(0) - yVolume(d.volume))
-    .attr('fill', '#9fc5e8');
-};
-
-// 선과 점 그리기 함수
-const drawLineAndPoints = (
-  svg: d3.Selection<SVGSVGElement, unknown, null, undefined>,
-  parsedData: ExtendedCandle[],
-  x: d3.ScaleBand<string>,
-  y: d3.ScaleLinear<number, number>,
-) => {
-  // 이전 데이터 clear
-  svg.selectAll('.line').remove();
-  svg.selectAll('circle').remove();
-
-  // 선 생성기
-  const line = d3
-    .line<ExtendedCandle>()
-    .x((d) => (x(d.dateStr) || 0) + x.bandwidth() / 2)
-    .y((d) => y(d.close));
-
-  // 선 그리기
-  svg
-    .append('path')
-    .datum(parsedData)
-    .attr('class', 'line')
-    .attr('fill', 'none')
-    .attr('stroke', 'steelblue')
-    .attr('stroke-width', 2)
-    .attr('d', line);
-
-  // 점 그리기
-  svg
-    .selectAll('circle')
-    .data(parsedData)
-    .join('circle')
-    .attr('cx', (d) => (x(d.dateStr) || 0) + x.bandwidth() / 2)
-    .attr('cy', (d) => y(d.close))
-    .attr('r', 1.5)
-    .attr('fill', 'tomato');
-};
-
-// 차트 렌더링 메인 함수
-export const renderChart = (
-  refinedData: RefinedCandle[],
-  svgRef: SVGSVGElement,
-  gx: SVGSVGElement,
-  gy: SVGSVGElement,
-  width: number,
-  height: number,
-  volumeHeight: number,
-  margin: { top: number; right: number; bottom: number; left: number },
-) => {
-  if (!refinedData || !refinedData.length || !svgRef) return;
-
-  // 그래프를 그릴 대상 dom
-  const svg = d3.select(svgRef);
-
-  // 데이터 파싱
-  const parsedData = parseDataWithDates(refinedData);
-  // 스케일 생성
-  const {
-    x,
-    y: yPrice,
-    yVolume,
-  } = createChartScales(parsedData, width, height, margin, volumeHeight);
-
-  // 축 그리기 (yPrice, yVolume 모두 넘김)
-  drawAxes(parsedData, x, yPrice, yVolume, gx, gy, width, height, margin);
-
-  // 그래프
-  drawLineAndPoints(svg, parsedData, x, yPrice);
-  drawVolumeBars(svg, parsedData, x, yVolume);
-};
